requirements-test.txt
pytest
pytest-asyncio
python-dotenv
rich